Update amazon-lightsail-using-route-53-to-point-a-domain-to-an-instance.md
Hello, 

I have deleted the following information and updated with the latest one. 

******* Deleted Section ********

**A – IPv4 address**  
An address \(A\) record maps a domain, such as `example.com`, or a subdomain, such as `blog.example.com`, to a web server’s IP address, such as `192.0.2.255`\.  

   1. Keep the **Name** text box empty to point the apex of your domain, such as `example.com`, to an IP address, or enter a subdomain\.

   1. Choose **A – IPv4 address** in the **Type** drop\-down menu\.

   1. Enter the static IP address \(public IP address\) of your Lightsail instance in the **Value** text box\.

   1. Keep the routing policy as **Simple**\.  
![\[Address record example in a Route 53 hosted zone.\]](https://d9yljz1nd5001.cloudfront.net/en_us/2f596334045058acdba2fdcc5e035cef/images/amazon-lightsail-route-53-hosted-zone-a-record.png)   

***************************************************************

Please review my changes and accept it changes if its relevant and useful and meet the requirements. 

  - Muhammad